CRA-297
Made pdf3dgaussian.cpp for handling of facies probabilities. 

git-svn-id: https://code.nr.no/scm/svn/crava/branches/StatoilDevelopment@1320 176de4d3-4747-0410-aa96-47f6d6634214

+++ rplib/trinormalwith2dtrend.h
@@ -82,7 +82,7 @@ class TriNormalWith2DTrend {
                          const double & obs_vp,
                          const double & obs_vs,
                          const double & obs_rho,
-                         float        & prob) const;
+                         double       & prob) const;
 
    double** DebugCreateEstimateOfCovMatrix(const NRLib::Normal & vp01,
                                            const NRLib::Normal & vs01,
